feat(#869): remove redundant class
