--- src/main/java/org/eolang/jeo/BytecodeDirectory.java
@@ -1,138 +1,0 @@
-/*
- * The MIT License (MIT)
- *
- * Copyright (c) 2016-2024 Objectionary.com
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included
- * in all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NON-INFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-package org.eolang.jeo;
-
-import com.jcabi.log.Logger;
-import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.util.List;
-import java.util.Optional;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
-import org.objectweb.asm.ClassReader;
-import org.objectweb.asm.Opcodes;
-import org.objectweb.asm.Type;
-import org.objectweb.asm.tree.ClassNode;
-import org.objectweb.asm.tree.MethodNode;
-import org.objectweb.asm.tree.analysis.Analyzer;
-import org.objectweb.asm.tree.analysis.AnalyzerException;
-import org.objectweb.asm.tree.analysis.SimpleVerifier;
-import org.objectweb.asm.util.CheckClassAdapter;
-
-/**
- * This class knows how to verify generated bytecode.
- * It requires all the classes to be loaded into the current classloader.
- * See {@link PluginStartup#init()} how to load all the generated classes.
- * @since 0.6
- */
-final class BytecodeDirectory {
-
-    /**
-     * Input directory where all the generated class files are placed.
-     */
-    private final Path input;
-
-    /**
-     * Constructor.
-     * @param input Input directory where all the generated class files are placed.
-     */
-    BytecodeDirectory(final Path input) {
-        this.input = input;
-    }
-
-    /**
-     * Verify bytecode in the folder.
-     */
-    void verify() {
-        try (
-            Stream<Path> classess = Files.walk(this.input)
-                .filter(Files::isRegularFile)
-                .filter(path -> path.toString().endsWith(".class"))
-        ) {
-            classess.map(BytecodeDirectory::read).forEach(BytecodeDirectory::verify);
-        } catch (final IOException exception) {
-            throw new IllegalStateException(
-                String.format("Can't read '%s' directory with classes", this.input),
-                exception
-            );
-        }
-    }
-
-    /**
-     * Read bytes of the class file.
-     * @param clazz Class file to read.
-     * @return Bytes of the class file.
-     */
-    private static byte[] read(final Path clazz) {
-        try {
-            return Files.readAllBytes(clazz);
-        } catch (final IOException exception) {
-            throw new IllegalStateException(
-                String.format("Can't read bytecode from the file '%s'", clazz),
-                exception
-            );
-        }
-    }
-
-    /**
-     * Verify the generated bytecode.
-     * @param bytes Bytecode to verify.
-     */
-    private static void verify(final byte[] bytes) {
-        final ClassNode clazz = new ClassNode();
-        new ClassReader(bytes)
-            .accept(new CheckClassAdapter(clazz, false), ClassReader.SKIP_DEBUG);
-        final Optional<Type> syper = Optional.ofNullable(clazz.superName)
-            .map(Type::getObjectType);
-        final List<Type> interfaces = clazz.interfaces.stream().map(Type::getObjectType)
-            .collect(Collectors.toList());
-        for (final MethodNode method : clazz.methods) {
-            try {
-                final SimpleVerifier verifier =
-                    new SimpleVerifier(
-                        Type.getObjectType(clazz.name),
-                        syper.orElse(null),
-                        interfaces,
-                        (clazz.access & Opcodes.ACC_INTERFACE) != 0
-                    );
-                verifier.setClassLoader(Thread.currentThread().getContextClassLoader());
-                new Analyzer<>(verifier).analyze(clazz.name, method);
-            } catch (final ClassFormatError | AnalyzerException exception) {
-                throw new IllegalStateException(
-                    String.format(
-                        "Bytecode verification failed for the class '%s' and method '%s'",
-                        clazz.name,
-                        method.name
-                    ),
-                    exception
-                );
-            }
-        }
-        Logger.info(
-            BytecodeDirectory.class,
-            String.format("Bytecode verification passed for the class '%s'", clazz.name)
-        );
-    }
-}
